requirements.txt
Flask==2.3.2
flask_jwt_extended==4.5.2 
Flask-PyMongo==2.3.0
bcrypt==3.2.2
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==3.0.1
Flask-Uploads==0.2.1
